+++ R/globals.R
@@ -1,5 +1,6 @@
 #'
 utils::globalVariables(c("name", "iso_a2", "iso_a3", "geometry",
-                         "MapFiller", "X", "Y",
-                         "iso_a2_eh", "iso_a3_eh",
-                         "map_df0"))
+                         "MapFiller", "X", "Y"#,
+                         #"iso_a2_eh", "iso_a3_eh",
+                         #"map_df0"
+                         ))