+++ R/worldplot.R
@@ -68,8 +68,8 @@ worldplot <- function(data,
   
   simdata$iso_a2 <- replace(simdata$iso_a2, simdata$iso_a2 == "UK", "GB")
 
-  map_df <- left_join(map_df0, simdata, by = "iso_a2")
-  
+  #map_df <- left_join(map_df0, simdata, by = "iso_a2")
+  map_df <- merge(WorldMapR::map_df0, simdata, by = "iso_a2", all.x = TRUE)
 
   if (missing(rangeVal)) {
     rangeVal = c(range(map_df$MapFiller, na.rm = TRUE))